feat: refactor code and add debugging utils
