+++ delay_service.c
@@ -1,7 +1,7 @@
 #include "delay_service.h"
 
 /* 
- * Deelay program through cycles.
+ * Delay program through cycles.
  * SystemCoreClock has a frequency of 20,971,520 Hz.
  * Setting duration to 0x0FFFFF = 1,048,575 delay 1/20 of a second.
 */
