+++ lab_service.h
@@ -1,0 +1,2 @@
+void lab_one(void);
+void lab_two(void);
