+++ switch_service.h
@@ -1,0 +1,3 @@
+#define SWITCH_BUTTON_PIN 6
+
+void switch_button_init(void);
