+++ utils.h
@@ -1,1 +1,3 @@
+#define UTILS_DEBUG_OUPUT_PIN 6
+
 #define MASK(x) (1 << (x))
